+++ docs/docs/users/reference/cli.sh
@@ -90,6 +90,7 @@ generate_markdown_section "forest-cli" "f3 powertable get-proportion"
 generate_markdown_section "forest-tool" ""
 
 generate_markdown_section "forest-tool" "backup"
+generate_markdown_section "forest-tool" "completion"
 generate_markdown_section "forest-tool" "backup create"
 generate_markdown_section "forest-tool" "backup restore"
 
@@ -141,3 +142,5 @@ generate_markdown_section "forest-tool" "shed summarize-tipsets"
 generate_markdown_section "forest-tool" "shed peer-id-from-key-pair"
 generate_markdown_section "forest-tool" "shed private-key-from-key-pair"
 generate_markdown_section "forest-tool" "shed openrpc"
+generate_markdown_section "forest-tool" "shed f3"
+generate_markdown_section "forest-tool" "shed f3 check-activation-raw"