+++ scripts/tests/calibnet_no_discovery_check.sh
@@ -12,6 +12,8 @@ FOREST_NODE_PID=$!
 FULLNODE_API_INFO="$(cat admin_token):/ip4/127.0.0.1/tcp/2345/http"
 export FULLNODE_API_INFO
 
+forest_wait_api
+
 # Verify that one of the seed nodes has been connected to
 until $FOREST_CLI_PATH net peers | grep "calib"; do
     sleep 1s;