+++ src/rpc/methods/gas.rs
@@ -163,7 +163,7 @@ pub async fn estimate_gas_premium<DB: Blockstore>(
     // mean 1, stddev 0.005 => 95% within +-1%
     let noise: f64 = Normal::new(1.0, 0.005)
         .unwrap()
-        .sample(&mut rand::thread_rng());
+        .sample(&mut crate::utils::rand::forest_rng());
 
     premium *= BigInt::from_f64(noise * (1i64 << precision) as f64)
         .context("failed to convert gas premium f64 to bigint")?;