+++ src/rpc/mod.rs
@@ -91,6 +91,7 @@ macro_rules! for_each_rpc_method {
         $callback!($crate::rpc::eth::EthGetBlockTransactionCountByNumber);
         $callback!($crate::rpc::eth::EthGetCode);
         $callback!($crate::rpc::eth::EthGetLogs);
+        $callback!($crate::rpc::eth::EthGetFilterLogs);
         $callback!($crate::rpc::eth::EthGetMessageCidByTransactionHash);
         $callback!($crate::rpc::eth::EthGetStorageAt);
         $callback!($crate::rpc::eth::EthGetTransactionByHash);
@@ -267,6 +268,7 @@ macro_rules! for_each_rpc_method {
         $callback!($crate::rpc::f3::Finalize);
         $callback!($crate::rpc::f3::ProtectPeer);
         $callback!($crate::rpc::f3::SignMessage);
+        $callback!($crate::rpc::f3::GetManifestFromContract);
 
         // misc
         $callback!($crate::rpc::misc::GetActorEventsRaw);